whitespace

+++ fyi_msg/src/msg.rs
@@ -885,7 +885,10 @@ impl Msg {
 				.fold(
 					self.toc.start(PART_TIMESTAMP) + 14,
 					|dst_off, x| {
-						ptr.add(dst_off).copy_from_nonoverlapping(utility::time_format_dd(*x).as_ptr(), 2);
+						ptr.add(dst_off).copy_from_nonoverlapping(
+							utility::time_format_dd(*x as usize).as_ptr(),
+							2
+						);
 						dst_off + 3
 					}
 				);
